requirements.txt
fastapi~=0.95.1
uvicorn~=0.22.0
sqlalchemy~=2.0.12
pydantic~=1.10.7
pydantic[email]
passlib[bcrypt]~=1.7.4
python-jose[cryptography]
python-multipart
jwt